+++ link-move/src/test/java/com/nhl/link/move/itest/runtime/task/createorupdate/DefaultCreateOrUpdateBuilderIT.java
@@ -4,9 +4,10 @@ import com.nhl.link.move.runtime.cayenne.ITargetCayenneService;
 import com.nhl.link.move.runtime.cayenne.TargetCayenneService;
 import com.nhl.link.move.runtime.extractor.IExtractorService;
 import com.nhl.link.move.runtime.task.MapperBuilder;
+import com.nhl.link.move.runtime.task.common.FkResolver;
+import com.nhl.link.move.runtime.task.createorupdate.CreateOrUpdateTargetMerger;
 import com.nhl.link.move.runtime.task.createorupdate.DefaultCreateOrUpdateBuilder;
 import com.nhl.link.move.runtime.task.createorupdate.RowConverter;
-import com.nhl.link.move.runtime.task.createorupdate.CreateOrUpdateTargetMerger;
 import com.nhl.link.move.runtime.token.ITokenManager;
 import com.nhl.link.move.unit.DerbySrcTargetTest;
 import com.nhl.link.move.unit.cayenne.t.Etl1t;
@@ -32,6 +33,7 @@ public class DefaultCreateOrUpdateBuilderIT extends DerbySrcTargetTest {
         DefaultCreateOrUpdateBuilder<Etl1t> builder = new DefaultCreateOrUpdateBuilder<>(
                 Etl1t.class,
                 mock(CreateOrUpdateTargetMerger.class),
+                mock(FkResolver.class),
                 mock(RowConverter.class),
                 cayenneService,
                 mock(IExtractorService.class),
@@ -46,6 +48,7 @@ public class DefaultCreateOrUpdateBuilderIT extends DerbySrcTargetTest {
         DefaultCreateOrUpdateBuilder<Etl1t> builder = new DefaultCreateOrUpdateBuilder<>(
                 Etl1t.class,
                 mock(CreateOrUpdateTargetMerger.class),
+                mock(FkResolver.class),
                 mock(RowConverter.class),
                 cayenneService,
                 mock(IExtractorService.class),