+++ link-move/src/test/java/com/nhl/link/move/unit/cayenne/t/Etl4t_jt.java
@@ -1,0 +1,9 @@
+package com.nhl.link.move.unit.cayenne.t;
+
+import com.nhl.link.move.unit.cayenne.t.auto._Etl4t_jt;
+
+public class Etl4t_jt extends _Etl4t_jt {
+
+    private static final long serialVersionUID = 1L; 
+
+}